# sdk/openapi.yaml
openapi: 3.0.3
info:
  title: DJ Tech Agent Runtime API
  version: 1.0.0
  description: |
    Public API for DJ Tech's multi-agent orchestration runtime.
    Create, manage, and execute AI agents with secure sandboxing, intelligent LLM routing,
    durable job queues, and usage-based billing.
  contact:
    name: DJ Tech
    url: https://djtech.io
    email: support@djtech.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.djtech.io/v1
    description: Production server
  - url: https://api-staging.djtech.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /agents:
    post:
      summary: Create a new agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List agents
      operationId: listAgents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, stopped, failed, deleted]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
...
  /billing/usage:
    get:
      summary: Get usage and billing information
      operationId: getUsage
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: agent_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, running, stopped, failed, deleted]
        security_tier:
          type: string
          enum: [standard, high, system]
        runtime:
          $ref: '#/components/schemas/AgentRuntime'
        resources:
          $ref: '#/components/schemas/ResourceLimits'
        llm:
          $ref: '#/components/schemas/LLMConfig'
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolConfig'
        networking:
          $ref: '#/components/schemas/NetworkConfig'
        autoscaling:
          $ref: '#/components/schemas/AutoscalingConfig'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        stopped_at:
          type: string
          format: date-time
    AgentRuntime:
      type: object
      properties:
        image:
          type: string
        entrypoint:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: string
    ResourceLimits:
      type: object
      properties:
        cpu:
          type: string
          description: CPU limit (e.g., "1000m" for 1 core)
        memory:
          type: string
          description: Memory limit (e.g., "2Gi")
        disk:
          type: string
          description: Disk limit (e.g., "1Gi")
        max_execution_time:
          type: string
          description: Max execution time (e.g., "24h")
    LLMConfig:
      type: object
      properties:
        default_model:
          type: string
        max_tokens_per_request:
          type: integer
        monthly_budget_usd:
          type: number
        allowed_models:
          type: array
          items:
            type: string
    ToolConfig:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [builtin, sandboxed, mcp]
        config:
          type: object
          additionalProperties: true
    NetworkConfig:
      type: object
      properties:
        egress_allowlist:
          type: array
          items:
            type: string
    AutoscalingConfig:
      type: object
      properties:
        min_replicas:
          type: integer
          minimum: 0
        max_replicas:
          type: integer
          minimum: 1
        scale_up_delay_seconds:
          type: integer
          minimum: 0
    CreateAgentRequest:
      type: object
      required:
        - name
        - runtime
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        security_tier:
          type: string
          enum: [standard, high, system]
          default: standard
        runtime:
          $ref: '#/components/schemas/AgentRuntime'
        resources:
          $ref: '#/components/schemas/ResourceLimits'
        llm:
          $ref: '#/components/schemas/LLMConfig'
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolConfig'
        networking:
          $ref: '#/components/schemas/NetworkConfig'
        autoscaling:
          $ref: '#/components/schemas/AutoscalingConfig'
...
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    UsageResponse:
      type: object
      properties:
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        total_cost_usd:
          type: number
        breakdown:
          type: array
          items:
            $ref: '#/components/schemas/UsageBreakdown'
    UsageBreakdown:
      type: object
      properties:
        agent_id:
          type: string
          format: uuid
        agent_name:
          type: string
        event_type:
          type: string
          enum: [llm_tokens, tool_call, sandbox_seconds, storage_bytes]
        quantity:
          type: integer
        unit_cost_usd:
          type: number
        total_cost_usd:
          type: number
...